requirements.txt
dtaidistance==2.3.10
matplotlib==3.5.3
mediapipe==0.9.0.1
numpy==1.20.3
open3d==0.17.0
opencv_contrib_python==3.4.11.45
opencv_python==4.7.0.72
opencv_python_headless==4.5.3.56
pandas==1.3.5
pyVHR==2.0
PyYAML==6.0.1
scikit_learn==1.0.2
scipy==1.7.3
seaborn==0.11.1
statsmodels==0.12.2
tqdm==4.61.1
~atplotlib==3.3.4
